+++ src/ios/APPLocalNotification.m
@@ -53,6 +53,8 @@
 - (NSString*) applicationState;
 // Retrieves all scheduled notifications
 - (NSArray*) scheduledNotifications;
+// Simply invokes the callback without any parameter
+- (void) execCallback:(CDVInvokedUrlCommand*)command;
 // Fires the given event
 - (void) fireEvent:(NSString*)event id:(NSString*)id json:(NSString*)json;
 
@@ -133,6 +135,8 @@
         if (notification) {
             [self cancelNotification:notification fireEvent:YES];
         }
+
+        [self execCallback:command];
     }];
 }
 
@@ -153,6 +157,8 @@
 
         [[UIApplication sharedApplication]
          setApplicationIconBadgeNumber:0];
+
+        [self execCallback:command];
     }];
 }
 
@@ -558,6 +564,18 @@
     }
 
     return notificationsWithoutNIL;
+}
+
+/**
+ * Simply invokes the callback without any parameter.
+ */
+- (void) execCallback:(CDVInvokedUrlCommand*)command
+{
+    CDVPluginResult *result = [CDVPluginResult
+                               resultWithStatus:CDVCommandStatus_OK];
+
+    [self.commandDelegate sendPluginResult:result
+                                callbackId:command.callbackId];
 }
 
 /**