+++ src/wp8/LocalNotification.cs
@@ -84,6 +84,7 @@ namespace Cordova.Extension.Commands
             cancelAll(jsonArgs);
 
             FireEvent("cancel", notificationID, "");
+            DispatchCommandResult();
         }
 
         /// <summary>